Time: 7 ms (23.79%), Space: 50.1 MB (31.49%) - LeetHub

+++ 4-median-of-two-sorted-arrays/4-median-of-two-sorted-arrays.java
@@ -15,7 +15,7 @@ class Solution {
             int x1= x-1;
             double z =Double.valueOf(merged[x1]);
             double d =Double.valueOf(merged[x]);
-            median =Double.valueOf((d+z)/2);
+            median = (d+z)/2;
         }
         else 
         {
